// fontend_web/src/features/lecturers/models/NhatKy.ts
export interface NhatKy {
  id: string | number
  tuan?: number
  tenDeTai?: string
  maSV?: string
  lop?: string
  idDeTai?: string | number
  hoTen?: string
  ngayBatDau?: string
  ngayKetThuc?: string
  trangThaiNhatKy?: string
  noiDung?: string | null
  fileUrl?: string | null
  nhanXet?: string | null
  raw?: any
}

export default NhatKy